moved all database functions into database parser
